scheduler error : File exists: '/etc/bunkerweb/plugins'

+++ start.sh
@@ -15,11 +15,14 @@ fi
 
 git pull
 sudo docker compose down
-sudo rm -rf $JOHNCLOUD_ROOT/bunkerweb/*
+if [[ "${ERASE_BW_FOLDERS}" = "yes" ]]; then
+  echo "Erasing $JOHNCLOUD_ROOT/bunkerweb/*"
+  sudo rm -rf $JOHNCLOUD_ROOT/bunkerweb/*
+fi
 sudo mkdir -p $JOHNCLOUD_ROOT/bunkerweb/bw-storage
 sudo chown -R root:101 $JOHNCLOUD_ROOT/bunkerweb/bw-storage
 sudo chmod -R 770 $JOHNCLOUD_ROOT/bunkerweb/bw-storage
-# sudo cp ./syslog-ng.conf $JOHNCLOUD_ROOT/bunkerweb/syslog-ng.conf
+sudo cp ./syslog-ng.conf $JOHNCLOUD_ROOT/bunkerweb/syslog-ng.conf
 sudo docker compose pull
 sudo docker compose up -d --remove-orphans --force-recreate
 sudo docker compose logs -f
